+++ pos-api-bff/src/schemas/salesSchemas.ts
@@ -2,12 +2,12 @@ import { z } from 'zod';
 
 export const validateStockSchema = z.object({
   productId: z.string().min(1),
-  quantity: z.number().int().positive(),
+  quantity: z.number().positive(),
 });
 
 export const createSaleSchema = z.object({
   productId: z.string().min(1),
-  quantity: z.number().int().positive(),
+  quantity: z.number().positive(),
   customerId: z.string().min(1),
   paymentMethod: z.enum(['cash', 'card', 'transfer']),
 });
